% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_split.R
\name{gd_split}
\alias{gd_split}
\title{Split a string with a separator}
\usage{
gd_split(text, split)
}
\arguments{
\item{text}{A character vector with one element}

\item{split}{Split based on}
}
\value{
Returns a vector of characters
}
\description{
Split a string with a separator
}
\examples{
text <- c("sui,ka,moku")
gd_split(text = text, split =",")
gd_split(text = c("sui,ka,moku"), split = ",")

}
