% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_title.R
\name{gd_title}
\alias{gd_title}
\title{Convert string to title case}
\usage{
gd_title(text)
}
\arguments{
\item{text}{A string which to be capitalized.}
}
\value{
A string with each word capitalized.
}
\description{
Convert string to title case
}
\examples{
x<-"a text"
gd_title(x)
gd_title("hello world!")

}
